SKILIKS-5015

+++ protected/commands/GenerateFullAssessmentAnalyticFileCommand.php
@@ -4,14 +4,19 @@
  */
 class GenerateFullAssessmentAnalyticFileCommand extends CConsoleCommand
 {
-    public function actionIndex()
+    public function actionIndex($email=null)
     {
-        $project_path = __DIR__."/../../";
-        $path = "rm -rf ".$project_path."protected/system_data/analytic_files_2/*";
-        echo "Start ".$path."\r\n";
-        exec($path);
-        echo "End\r\n";
-        $users_account = UserAccountCorporate::model()->findAll();
+        if($email !== null){
+            $profile = YumProfile::model()->findAllByAttributes(['email'=>$email]);
+            $users_account = UserAccountCorporate::model()->findAllByAttributes(['user_id'=>$profile->user_id]);
+        } else {
+            $project_path = __DIR__."/../../";
+            $path = "rm -rf ".$project_path."protected/system_data/analytic_files_2/*";
+            echo "Start ".$path."\r\n";
+            exec($path);
+            echo "End\r\n";
+            $users_account = UserAccountCorporate::model()->findAll();
+        }
         /* @var UserAccountCorporate[] $users_account */
         echo "Found ".count($users_account)." accounts \n";
         foreach($users_account as $account) {
